Chatbot/requirements.txt
-U pip setuptools wheel
-U spacy[cuda111]
botbuilder-integration-aiohttp>=4.11.0
